# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carmarket API",
        "contact": {
            "name": "Support Team",
            "email": "support@ticketzone.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/inside/vehicle": {
            "post": {
                "summary": "Store a single vehicle from Inside",
                "description": "Store vehicle from Inside",
                "operationId": "5c57b74edef6e71bef5bd34dcc7e7a50",
                "requestBody": {
                    "description": "Request body for vehicle store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsideVehicleStoreRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessMessageSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AuthJWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationSchema": {
                "required": [
                    "error_key",
                    "error_message",
                    "code",
                    "errors"
                ],
                "properties": {
                    "error": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string"
                                },
                                "field": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SuccessMessageSchema": {
                "required": [
                    "stauts",
                    "message"
                ],
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "AuthJWT": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}